--- a/app/jobs/destroy_problems_by_app_job.rb
+++ b/app/jobs/destroy_problems_by_app_job.rb
@@ -3,6 +3,7 @@
 class DestroyProblemsByAppJob < ActiveJob::Base
   queue_as :default
 
+  # @param app_id [String]
   def perform(app_id)
     app = App.find_by(id: app_id)