Before:

```
rspec spec/jobs/destroy_problems_by_app_job_spec.rb --profile

Randomized with seed 45881

DestroyProblemsByAppJob
  destroys all problems, even with a large amount of them
  should work with a fresh new application
  destroys all problems

Top 3 slowest examples (23.47 seconds, 100.0% of total time):
  DestroyProblemsByAppJob destroys all problems, even with a large amount of them
    23.34 seconds ./spec/jobs/destroy_problems_by_app_job_spec.rb:19
  DestroyProblemsByAppJob should work with a fresh new application
    0.06987 seconds ./spec/jobs/destroy_problems_by_app_job_spec.rb:30
  DestroyProblemsByAppJob destroys all problems
    0.05931 seconds ./spec/jobs/destroy_problems_by_app_job_spec.rb:12

Finished in 23.47 seconds (files took 1.98 seconds to load)
3 examples, 0 failures
```

After:

```
rspec spec/jobs/destroy_problems_by_app_job_spec.rb --profile

Randomized with seed 57248

DestroyProblemsByAppJob
  should work with a fresh new application
  destroys all problems

Top 2 slowest examples (0.26001 seconds, 99.0% of total time):
  DestroyProblemsByAppJob should work with a fresh new application
    0.19385 seconds ./spec/jobs/destroy_problems_by_app_job_spec.rb:17
  DestroyProblemsByAppJob destroys all problems
    0.06616 seconds ./spec/jobs/destroy_problems_by_app_job_spec.rb:6

Finished in 0.26272 seconds (files took 4.23 seconds to load)
2 examples, 0 failures
```

Total: minus 23 seconds